backend/requirements.txt
# Core RAG retrieval and QA
faiss-cpu==1.13.2
numpy==2.4.2
rank-bm25==0.2.2
sentence-transformers==5.2.3
torch==2.10.0
transformers==5.3.0

# API and UI
fastapi==0.135.1
gradio==6.9.0
gradio_client==2.3.0
uvicorn==0.41.0

# Data handling and ML utilities
pandas==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
tqdm==4.67.3
requests==2.32.5

# Optional OCR/image tooling used in ai_module
easyocr==1.7.2
opencv-python-headless==4.13.0.92
pillow==12.1.1